requirements.txt
streamlit==1.27.2
numpy==1.25.2
scipy==1.11.2
pandas==2.1.0
scikit-learn==1.3.0
supabase==1.0.3
pydantic==2.4.2
pytest==7.4.2
python-dotenv==1.0.0
